feat: add Handleliste
This is a start, where the shopping list has the main components for the
MCU. We also need to buy a debug pin thing, and probably a connector
which goes into the debug pin header.

We also need a USB-A port.

At a later stage we need to include all the FPGA-related components and
the power supply.

I am not looking forward to find all the specific capacitors&resistors we need to
buy :(((