feat: add power_down_transformator symbol
I've moved this symbol to the *tegnebrett_symbols* file.
Apparently it is the case that multiple symbols are stored within one
.kicad_sym file.

I **also** moved the PJ (Power Jack) symbol to this file.
Preparing to remove the PJ symbol file, and also the symbols folder.
Everything is stored in the (misnamed) USB....kicad_sym file.