feat: add buttons and VGA transceiver to shopping list
We had forgotten to add these elements to the list. Now the list
consists of 211 components per PCB.

The Bill of Materials (BoM) is telling us that we have footprints for
203 components per PCB.
We get 211 components per PCB, because 8 components do not have a
footprint on the board. These components are:

LCD Display (1x)
LCD Breakout 45pin (1x)
LCD Breakout 6pin (1x)
15V Power Supply (1x)
Jumper Cap (4x)