fix: change *zone connection style* for thermal vias
Change from *inherited* to *solid*.
This means that there are no longer any *copper spokes* going from the
thermal vias to the rest of the thermal pad (which also has pin number
17).

This *increases* thermal conductivity, but makes it somewhat *harder* to
solder the chip onto the PCB when we're doing hand-soldering.

Crucially it removes errors in the PCB layout when applied.
After this footprint update, errors in the PCB layout drop to 0.